healthcare-knowledge-rag-assistant/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
annotated-doc==0.0.4
    # via
    #   fastapi
    #   typer
annotated-types==0.7.0
    # via pydantic
anyio==4.12.1
    # via
    #   httpx
    #   starlette
certifi==2026.2.25
    # via
    #   httpcore
    #   httpx
click==8.3.1
    # via
    #   typer
    #   uvicorn
cuda-bindings==12.9.4
    # via torch
cuda-pathfinder==1.4.0
    # via cuda-bindings
faiss-cpu==1.13.2
    # via -r requirements.in
fastapi==0.133.1
    # via -r requirements.in
filelock==3.24.3
    # via
    #   huggingface-hub
    #   torch
fsspec==2026.2.0
    # via
    #   huggingface-hub
    #   torch
googletrans==4.0.2
    # via -r requirements.in
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
h2==4.3.0
    # via httpx
hf-xet==1.3.1
    # via huggingface-hub
hpack==4.1.0
    # via h2
httpcore==1.0.9
    # via httpx
httpx[http2]==0.28.1
    # via
    #   googletrans
    #   huggingface-hub
huggingface-hub==1.5.0
    # via
    #   sentence-transformers
    #   tokenizers
    #   transformers
hyperframe==6.1.0
    # via h2
idna==3.11
    # via
    #   anyio
    #   httpx
jinja2==3.1.6
    # via torch
joblib==1.5.3
    # via scikit-learn
langdetect==1.0.9
    # via -r requirements.in
markdown-it-py==4.0.0
    # via rich
markupsafe==3.0.3
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
mpmath==1.3.0
    # via sympy
networkx==3.6.1
    # via torch
numpy==2.4.2
    # via
    #   faiss-cpu
    #   scikit-learn
    #   scipy
    #   sentence-transformers
    #   transformers
nvidia-cublas-cu12==12.8.4.1
    # via
    #   nvidia-cudnn-cu12
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cuda-cupti-cu12==12.8.90
    # via torch
nvidia-cuda-nvrtc-cu12==12.8.93
    # via torch
nvidia-cuda-runtime-cu12==12.8.90
    # via torch
nvidia-cudnn-cu12==9.10.2.21
    # via torch
nvidia-cufft-cu12==11.3.3.83
    # via torch
nvidia-cufile-cu12==1.13.1.3
    # via torch
nvidia-curand-cu12==10.3.9.90
    # via torch
nvidia-cusolver-cu12==11.7.3.90
    # via torch
nvidia-cusparse-cu12==12.5.8.93
    # via
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cusparselt-cu12==0.7.1
    # via torch
nvidia-nccl-cu12==2.27.5
    # via torch
nvidia-nvjitlink-cu12==12.8.93
    # via
    #   nvidia-cufft-cu12
    #   nvidia-cusolver-cu12
    #   nvidia-cusparse-cu12
    #   torch
nvidia-nvshmem-cu12==3.4.5
    # via torch
nvidia-nvtx-cu12==12.8.90
    # via torch
packaging==26.0
    # via
    #   faiss-cpu
    #   huggingface-hub
    #   transformers
pydantic==2.12.5
    # via fastapi
pydantic-settings==2.13.1
    # required for BaseSettings (pydantic 2+)
pydantic-core==2.41.5
    # via pydantic
pygments==2.19.2
    # via rich
python-multipart==0.0.22
    # via -r requirements.in
pyyaml==6.0.3
    # via
    #   huggingface-hub
    #   transformers
regex==2026.2.19
    # via transformers
rich==14.3.3
    # via typer
safetensors==0.7.0
    # via transformers
scikit-learn==1.8.0
    # via sentence-transformers
scipy==1.17.1
    # via
    #   scikit-learn
    #   sentence-transformers
sentence-transformers==5.2.3
    # via -r requirements.in
shellingham==1.5.4
    # via typer
six==1.17.0
    # via langdetect
starlette==0.52.1
    # via fastapi
sympy==1.14.0
    # via torch
threadpoolctl==3.6.0
    # via scikit-learn
tokenizers==0.22.2
    # via transformers
torch==2.10.0
    # via sentence-transformers
tqdm==4.67.3
    # via
    #   huggingface-hub
    #   sentence-transformers
    #   transformers
transformers==5.2.0
    # via sentence-transformers
triton==3.6.0
    # via torch
typer==0.24.1
    # via
    #   huggingface-hub
    #   typer-slim
typer-slim==0.24.0
    # via transformers
typing-extensions==4.15.0
    # via
    #   anyio
    #   fastapi
    #   huggingface-hub
    #   pydantic
    #   pydantic-core
    #   sentence-transformers
    #   starlette
    #   torch
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   fastapi
    #   pydantic
uvicorn==0.41.0
    # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
